pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-security>=5.6.0",
    "python-dotenv>=1.0.1",
    "pyotp>=2.9.0",
    "flask-wtf>=1.2.2",
    "qrcode>=8.0",
    "pillow>=11.1.0",
    "flask-oauthlib>=0.9.6",
    "flask-migrate>=4.1.0",
    "authlib>=1.5.1",
]
